feat: migrate ecs api to new trpc api compute type
