+++ docs/src/content/docs/en/guides/ecs-trpc.mdx
@@ -14,7 +14,7 @@ import Snippet from '@components/snippet.astro';
 The ECS tRPC API generator creates a new tRPC API deployed as a containerised [AWS ECS Fargate](https://docs.aws.amazon.com/AmazonECS/latest/developerguide/AWS_Fargate.html) service, fronted by an [API Gateway HTTP API](https://docs.aws.amazon.com/apigateway/latest/developerguide/http-api.html) via a VPC Link. The generated backend uses [Fastify](https://fastify.dev/) as the HTTP server and includes schema validation using [Zod](https://zod.dev/).
 
 :::tip
-ECS Fargate is a good fit when your API has long-running requests, needs consistent low-latency responses without cold starts, or has steady traffic patterns that benefit from always-on compute. If your API is lightweight, has variable or bursty traffic, and can tolerate occasional cold starts, a serverless Lambda deployment may be simpler and more cost-effective. See the <Link path="/guides/trpc">`ts#trpc-api`</Link> generator for the Lambda-based alternative.
+ECS Fargate is a good fit when your API has long-running requests, needs consistent low-latency responses without cold starts, or has steady traffic patterns that benefit from always-on compute. If your API is lightweight, has variable or bursty traffic, and can tolerate occasional cold starts, a serverless Lambda deployment may be simpler and more cost-effective. See the <Link path="/guides/trpc">`ts#trpc-api` Lambda guide</Link> for the Lambda-based alternative.
 :::
 
 ## Usage
@@ -23,11 +23,13 @@ ECS Fargate is a good fit when your API has long-running requests, needs consist
 
 You can generate a new ECS tRPC API in two ways:
 
-<RunGenerator generator="ts#ecs-trpc-api" />
+<RunGenerator generator="ts#trpc-api" />
+
+Select `EcsFargate` as the `computeType` when prompted.
 
 ### Options
 
-<GeneratorParameters generator="ts#ecs-trpc-api" />
+<GeneratorParameters generator="ts#trpc-api" />
 
 ## Generator Output
