+++ docs/src/content/docs/en/guides/trpc.mdx
@@ -36,6 +36,10 @@ Select `ServerlessApiGatewayRestApi` (default) as your `computeType` if you woul
 The `integrationPattern` option defaults to `isolated`, which creates one Lambda per tRPC procedure. Select `shared` if you would prefer a single shared Lambda handler for the whole API, with optional per-procedure overrides.
 :::
 
+:::tip
+Select `EcsFargate` as your `computeType` to deploy your API as a containerised ECS Fargate service instead of Lambda. This is a good fit for long-running requests, consistent low-latency without cold starts, or steady traffic patterns. See the <Link path="/guides/ecs-trpc">ECS tRPC guide</Link> for details.
+:::
+
 ## Generator Output
 
 The generator will create the following project structure in the `<directory>/<api-name>` directory:
