+++ packages/nx-plugin/generators.json
@@ -241,12 +241,6 @@
       "description": "provides React integration to a React website",
       "metric": "g10",
       "hidden": true
-    },
-    "ts#ecs-trpc-api": {
-      "factory": "./src/ecs/trpc/generator",
-      "schema": "./src/ecs/trpc/schema.json",
-      "description": "Create an ECS Fargate API using tRPC and Fastify",
-      "metric": "g32"
     }
   }
 }
