--- packages/nx-plugin/src/ecs/trpc/schema.json
@@ -1,42 +1,0 @@
-{
-  "$schema": "https://json-schema.org/schema",
-  "$id": "EcsTrpcApi",
-  "title": "",
-  "type": "object",
-  "properties": {
-    "name": {
-      "type": "string",
-      "description": "The name of the API (required). Used to generate class names and file paths.",
-      "$default": {
-        "$source": "argv",
-        "index": 0
-      },
-      "x-prompt": "What name would you like your API to have? i.e: MyApi"
-    },
-    "auth": {
-      "type": "string",
-      "description": "The method used to authenticate with your API. Choose between IAM (default), Cognito or None.",
-      "default": "IAM",
-      "enum": ["IAM", "Cognito", "None"],
-      "x-prompt": "How would you like users to authenticate with your API?",
-      "x-priority": "important"
-    },
-    "directory": {
-      "description": "The directory to store the application in.",
-      "type": "string",
-      "alias": "dir",
-      "x-priority": "important",
-      "default": "packages",
-      "x-prompt": "What directory would you like to store your application in?"
-    },
-    "iacProvider": {
-      "type": "string",
-      "description": "The preferred IaC provider. By default this is inherited from your initial selection.",
-      "enum": ["Inherit", "CDK"],
-      "x-priority": "important",
-      "default": "Inherit",
-      "x-prompt": "Which provider would you like to manage your infrastructure? (default: Inherit)"
-    }
-  },
-  "required": ["name"]
-}
