+++ packages/nx-plugin/src/trpc/backend/files-ecs/Dockerfile.template
