+++ packages/nx-plugin/src/trpc/backend/files-ecs/src/index.ts.template
