+++ packages/nx-plugin/src/trpc/backend/files-ecs/src/init.ts.template
