+++ packages/nx-plugin/src/trpc/backend/files-ecs/src/local-server.ts.template
