+++ packages/nx-plugin/src/trpc/backend/files-ecs/src/server.ts.template
