+++ packages/nx-plugin/src/trpc/backend/schema.json
@@ -15,9 +15,13 @@
     },
     "computeType": {
       "type": "string",
-      "description": "The type of compute to use to deploy this API. Choose between ServerlessApiGatewayRestApi (default) or ServerlessApiGatewayHttpApi.",
+      "description": "The type of compute to use to deploy this API. Choose between ServerlessApiGatewayRestApi (default), ServerlessApiGatewayHttpApi, or EcsFargate.",
       "default": "ServerlessApiGatewayRestApi",
-      "enum": ["ServerlessApiGatewayRestApi", "ServerlessApiGatewayHttpApi"],
+      "enum": [
+        "ServerlessApiGatewayRestApi",
+        "ServerlessApiGatewayHttpApi",
+        "EcsFargate"
+      ],
       "x-prompt": "What compute type would you like to deploy your API with?",
       "x-priority": "important"
     },
@@ -66,7 +70,7 @@
     },
     "iacProvider": {
       "type": "string",
-      "description": "The preferred IaC provider. By default this is inherited from your initial selection.",
+      "description": "The preferred IaC provider. By default this is inherited from your initial selection. EcsFargate only supports CDK.",
       "enum": ["Inherit", "CDK", "Terraform"],
       "x-priority": "important",
       "default": "Inherit",
