+++ packages/nx-plugin/src/trpc/react/generator.ts
@@ -46,7 +46,7 @@ export async function reactGenerator(
   const auth = metadata.auth ?? 'IAM';
   const port = metadata.port ?? metadata.ports?.[0] ?? 2022;
   const isRestApi = metadata.computeType === 'ServerlessApiGatewayRestApi';
-  const isEcs = metadata.apiType === 'ecs-trpc';
+  const isEcs = metadata.computeType === 'EcsFargate';
   const apiNameClassName = toClassName(apiName);
   const backendProjectAlias = toScopeAlias(backendProjectConfig.name);
